Add files via upload
Most of my changes were just the creation of the scripts for each page, but I also edited the html files to include the appropriate script, the cordova script, and to create the 'results' div for the output to go into.  I also moved the script that controlled opening the boxes to be appended by my script so that it wouldn't bind those events until the boxes had been created.